First attempt at problem 3.
Uses a modified version of the Sieve of Eratosthenes to filter out all
of the prime numbers which are factors of the target number. This
approach seems to work for small target numbers, but fails with the
actual target because it uses too much memory and takes too long. A
different approach will be needed.